refactor(core): colocate serialized types with serialization logic
